4.21

+++ part4/blog/controllers/blogsRouter.js
@@ -42,6 +42,18 @@ blogsRouter.post('/', async (req, res) => {
 })
 
 blogsRouter.delete('/:id', async (req, res) => {
+	const token = req.token
+	const decodedToken = jwt.verify(token, process.env.SECRET)
+	if (!token || !decodedToken.id) {
+		return res.status(401).json({ error: 'token missing or invalid' })
+	}
+
+	const user = await User.findById(decodedToken.id)
+	const blog = await Blog.findById(req.params.id)
+	if (blog.user.toString() !== user._id.toString()) {
+		return res.status(401).json({ error: 'user not authorized to delete' })
+	}
+
 	await Blog.findByIdAndRemove(req.params.id)
 	res.status(204).end()
 })
